last versión

+++ database/migrations/2023_10_10_231829_create_inventario_materials_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreign('materiaId')->references('id')->on('materia_primas');
             $table->unsignedBigInteger('inventarioId');
             $table->foreign('inventarioId')->references('id')->on('inventarios');
-            $table->integer('cantidad');
+            $table->double('cantidad');
             $table->timestamps();
         });
     }
